+++ lib/api/bigcommerce/content.js
@@ -1,7 +1,7 @@
 import Handlebars from 'handlebars'
-import ifHelper from '../../if-helper'
-import unlessHelper from '../../unless-helper'
-import jsonHelper from '../../json-helper'
+import ifHelper from '../if-helper'
+import unlessHelper from '../unless-helper'
+import jsonHelper from '../json-helper'
 import fetchApi, { getErrorText, BigComerceError } from './fetch-api'
 
 export async function getHomepageContent() {
